Remove useAttendanceConnection flag from update()
In `CRUDLateAttendance` no need for `useAttendanceConnection` flag
because we always update the `Late` record with the presence of
a corresponding stored `Attendance` entity. This flag is useful only if
we need to commit changes from both `Attendance` and `Late` as atomic,
which does not happen at all in this case.