Switch the ability of the create button on and off
After computing salary for subject date, later any change in the
subject date, it means to switch off the ability of the create button.
If subject date is returned back to the subject date in which the last
salary computing was done then it is fine to switch on the ability of
the create button back.